requirements.txt
anthropic>=0.45.0
litellm>=1.50.0
fastapi>=0.115.0
uvicorn>=0.30.0
asyncpg>=0.30.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
aioboto3>=12.0.0
structlog>=24.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
